Pass min contig length to QUAST
Quast uses 500 by default, which may cause issues if the user specifies an MCL
less than that.

There are very few tests in the test suite and none at all for the QUAST output,
so I don't propose to spend a massive amount of time adding them. I checked
the KBaseReports for both tests
manually and confirmed that QUAST got the correct MCL.